add RVC option to pipeline
